requirements.txt
Django==5.2.6
djangorestframework==3.14.0
django-cors-headers==4.4.0
django-environ
python-dotenv
django-rest-framework
django-cors-headers
